List files in a directory

List files in a directory

dir_list(path = ".", all = FALSE, recursive = FALSE, type = "any",
  regexp = NULL, glob = NULL, ...)

dir_walk(path = ".", fun, all = FALSE, recursive = FALSE, type = "any")

dir_info(path = ".", all = FALSE, recursive = FALSE, type = "any",
  regexp = NULL, glob = NULL, ...)

Arguments

path
A character vector of paths.

all
If TRUE hidden files are also returned.

recursive
Should directories be listed recursively? the filenames.

type
File type(s) to return, one or more of "any", "file", "directory", "symlink", "FIFO", "socket", "character_device" or "block_device".

glob, regexp
Either a file glob (e.g. *.csv) or a regular expression (e.g. \.csv$) passed on to grep to filter paths.

...
Additional arguments passed to grep.

fun
A function, taking one parameter, the current path entry.

Functions

dir_walk: Walk along a directory, calling f for each entry in the directory.

dir_info: A shortcut for the combination of file_info(dir_list()).

Examples

dir_list(R.home("share/R"))

#> /Library/Frameworks/R.framework/Resources/share/R/REMOVE.R
#> /Library/Frameworks/R.framework/Resources/share/R/examples-footer.R
#> /Library/Frameworks/R.framework/Resources/share/R/examples-header.R
#> /Library/Frameworks/R.framework/Resources/share/R/nspackloader.R
#> /Library/Frameworks/R.framework/Resources/share/R/tests-startup.R

link_create(system.file(package = "base"), "base")

dir_list("base", recursive = TRUE, glob = "*.R")

#> base/demo/error.catching.R base/demo/is.things.R      
#> base/demo/recursion.R      base/demo/scoping.R        

dir_walk(system.file(), str)

#>  chr "/Library/Frameworks/R.framework/Resources/library/base/CITATION"
#>  chr "/Library/Frameworks/R.framework/Resources/library/base/DESCRIPTION"
#>  chr "/Library/Frameworks/R.framework/Resources/library/base/INDEX"
#>  chr "/Library/Frameworks/R.framework/Resources/library/base/Meta"
#>  chr "/Library/Frameworks/R.framework/Resources/library/base/R"
#>  chr "/Library/Frameworks/R.framework/Resources/library/base/demo"
#>  chr "/Library/Frameworks/R.framework/Resources/library/base/help"
#>  chr "/Library/Frameworks/R.framework/Resources/library/base/html"

dir_info("base")

#> # A tibble: 8 x 18
#>   path             type         size permissio… modification_time   user  group
#>   <fs::filename>   <fctr>  <fs::byt> <fs::perm> <dttm>              <chr> <chr>
#> 1 base/CITATION    file          956 rw-rw-r--  2017-12-06 21:16:12 root  admin
#> 2 base/DESCRIPTION file          286 rw-rw-r--  2017-12-06 21:16:12 root  admin
#> 3 base/INDEX       file        23.2K rw-rw-r--  2017-12-06 21:16:12 root  admin
#> 4 base/Meta        direct…       272 rwxrwxr-x  2017-12-06 21:16:12 root  admin
#> 5 base/R           direct…       204 rwxrwxr-x  2017-12-06 21:16:12 root  admin
#> 6 base/demo        direct…       204 rwxrwxr-x  2017-12-06 21:16:12 root  admin
#> 7 base/help        direct…       238 rwxrwxr-x  2017-12-06 21:16:12 root  admin
#> 8 base/html        direct…       136 rwxrwxr-x  2017-12-06 21:16:12 root  admin
#> # ... with 11 more variables: device_id <dbl>, hard_links <dbl>,
#> #   special_device_id <dbl>, inode <dbl>, block_size <dbl>, blocks <dbl>, flags
#> #   <int>, generation <dbl>, access_time <dttm>, creation_time <dttm>,
#> #   birth_time <dttm>

link_delete("base")