Delete files, directories, or links

file_delete() and link_delete() delete file and links. Compared to file.remove they always fail if they cannot delete the object rather than changing return value or signalling a warning. dir_delete() will first delete the contents of the directory, then remove the directory. Compared to unlink it will always throw an error if the directory cannot be deleted rather than being silent or signalling a warning.

file_delete(path)

dir_delete(path)

link_delete(path)

Arguments

path
A character vector of one or more paths.

Value

The deleted paths (invisibly).

Examples

# create a directory, with some files and a link to it
dir <- dir_create(file_temp())
files <- file_create(path(dir, letters[1:5]))
link <- link_create(path_abs(dir), "link")

# All files created
dir_exists(dir)

#> /tmp/filedd4635bd7c86 
#>                  TRUE 

file_exists(files)

#> /tmp/filedd4635bd7c86/a /tmp/filedd4635bd7c86/b /tmp/filedd4635bd7c86/c 
#>                    TRUE                    TRUE                    TRUE 
#> /tmp/filedd4635bd7c86/d /tmp/filedd4635bd7c86/e 
#>                    TRUE                    TRUE 

link_exists(link)

#> link 
#> TRUE 

file_exists(link_path(link))

#> /tmp/filedd4635bd7c86 
#>                  TRUE 

# Delete a file
file_delete(files[1])
file_exists(files[1])

#> /tmp/filedd4635bd7c86/a 
#>                   FALSE 

# Delete the directory (which deletes the files as well)
dir_delete(dir)
file_exists(files)

#> /tmp/filedd4635bd7c86/a /tmp/filedd4635bd7c86/b /tmp/filedd4635bd7c86/c 
#>                   FALSE                   FALSE                   FALSE 
#> /tmp/filedd4635bd7c86/d /tmp/filedd4635bd7c86/e 
#>                   FALSE                   FALSE 

dir_exists(dir)

#> /tmp/filedd4635bd7c86 
#>                 FALSE 

# The link still exists, but what it points to does not.
link_exists(link)

#> link 
#> TRUE 

dir_exists(link_path(link))

#> /tmp/filedd4635bd7c86 
#>                 FALSE 

# Delete the link
link_delete(link)
link_exists(link)

#>  link 
#> FALSE