List files in a directory

List files in a directory

dir_list(path = ".", all = FALSE, recursive = TRUE, type = "any",
  pattern = NULL, ...)

dir_walk(path = ".", fun, all = FALSE, recursive = TRUE, type = "any",
  pattern = NULL, ...)

Arguments

path
A character vector of paths.

all
If TRUE hidden files are also returned.

recursive
Should directories be listed recursively?

type
File type(s) to return, one or more of "any", "file", "directory", "symlink", "FIFO", "socket", "character_device" or "block_device".

pattern
A regular expression pattern used passed to grep to filter the filenames.

...
Additional arguments passed to grep.

fun
A function, taking one parameter, the current path entry.

Functions

dir_walk: Walk along a directory, calling f for each entry in the directory.

Examples

dir_list(system.file())

#>  [1] "/Library/Frameworks/R.framework/Resources/library/base/CITATION"             
#>  [2] "/Library/Frameworks/R.framework/Resources/library/base/DESCRIPTION"          
#>  [3] "/Library/Frameworks/R.framework/Resources/library/base/INDEX"                
#>  [4] "/Library/Frameworks/R.framework/Resources/library/base/Meta"                 
#>  [5] "/Library/Frameworks/R.framework/Resources/library/base/Meta/Rd.rds"          
#>  [6] "/Library/Frameworks/R.framework/Resources/library/base/Meta/demo.rds"        
#>  [7] "/Library/Frameworks/R.framework/Resources/library/base/Meta/features.rds"    
#>  [8] "/Library/Frameworks/R.framework/Resources/library/base/Meta/hsearch.rds"     
#>  [9] "/Library/Frameworks/R.framework/Resources/library/base/Meta/links.rds"       
#> [10] "/Library/Frameworks/R.framework/Resources/library/base/Meta/package.rds"     
#> [11] "/Library/Frameworks/R.framework/Resources/library/base/R"                    
#> [12] "/Library/Frameworks/R.framework/Resources/library/base/R/Rprofile"           
#> [13] "/Library/Frameworks/R.framework/Resources/library/base/R/base"               
#> [14] "/Library/Frameworks/R.framework/Resources/library/base/R/base.rdb"           
#> [15] "/Library/Frameworks/R.framework/Resources/library/base/R/base.rdx"           
#> [16] "/Library/Frameworks/R.framework/Resources/library/base/demo"                 
#> [17] "/Library/Frameworks/R.framework/Resources/library/base/demo/error.catching.R"
#> [18] "/Library/Frameworks/R.framework/Resources/library/base/demo/is.things.R"     
#> [19] "/Library/Frameworks/R.framework/Resources/library/base/demo/recursion.R"     
#> [20] "/Library/Frameworks/R.framework/Resources/library/base/demo/scoping.R"       
#> [21] "/Library/Frameworks/R.framework/Resources/library/base/help"                 
#> [22] "/Library/Frameworks/R.framework/Resources/library/base/help/AnIndex"         
#> [23] "/Library/Frameworks/R.framework/Resources/library/base/help/aliases.rds"     
#> [24] "/Library/Frameworks/R.framework/Resources/library/base/help/base.rdb"        
#> [25] "/Library/Frameworks/R.framework/Resources/library/base/help/base.rdx"        
#> [26] "/Library/Frameworks/R.framework/Resources/library/base/help/paths.rds"       
#> [27] "/Library/Frameworks/R.framework/Resources/library/base/html"                 
#> [28] "/Library/Frameworks/R.framework/Resources/library/base/html/00Index.html"    
#> [29] "/Library/Frameworks/R.framework/Resources/library/base/html/R.css"           

dir_walk(system.file(), function(p) if (grepl("profile", p)) print(p))

#> [1] "/Library/Frameworks/R.framework/Resources/library/base/R/Rprofile"