file_chown {fs}	R Documentation

Change owner or group of a file

Description

Change owner or group of a file

Usage

file_chown(path, user_id = NULL, group_id = NULL)

Arguments

path
A character vector of one or more paths.

user_id
The user id of the new owner, specified as a numeric ID or name. The R process must be privileged to change this.

group_id
The group id of the new owner, specified as a numeric ID or name.

[Package fs version 1.6.4 Index]