file_touch {fs}	R Documentation

Change file access and modification times

Description

Unlike the touch POSIX utility this does not create the file if it does not exist. Use file_create() to do this if needed.

Usage

file_touch(path, access_time = Sys.time(), modification_time = access_time)

Arguments

path
A character vector of one or more paths.

access_time, modification_time
The times to set, inputs will be coerced to POSIXct objects.

Examples

file_create("foo")
file_touch("foo", "2018-01-01")
file_info("foo")[c("access_time", "modification_time", "change_time", "birth_time")]

[Package fs version 1.5.2 Index]